test: Refactor codemod test cases for docstring and parameter updates
