+++ tests/codemods/codemod_test_cases/class_attr_add_no_docstr/before.py
