+++ tests/codemods/codemod_test_cases/func_docstring_add/before.py
@@ -1,11 +1,11 @@
 """
-simple functions 
+Add docstring if not provided
 """
 
 
-def foo():
+def foo(self, iterable) -> None:
+    # no docstring
     ...
 
 
-def bar():
-    ...
+def bar(): ...
