+++ tests/codemods/codemod_test_cases/func_docstring_add_no_params/before.py
@@ -3,7 +3,7 @@ simple functions
 """
 
 
-def foo(x=1, y=2):
+def foo():
     ...
