+++ tests/codemods/codemod_test_cases/func_docstring_add_no_params/doc_stub.py
@@ -1,0 +1,10 @@
+"""
+simple functions 
+"""
+
+
+def foo(pin: int, /, limit: int = 100) -> str:
+    """
+    simple docstring
+    """
+    ...
