+++ tests/codemods/codemod_test_cases/func_docstring_add_no_params/expected.py
@@ -4,4 +4,11 @@ simple functions
 
 
 def foo(pin: int, /, limit: int = 100) -> str:
+    """
+    simple docstring
+    """
+    ...
+
+
+def bar():
     ...
