+++ tests/codemods/codemod_test_cases/func_docstring_no_overwrite_no_docstr/doc_stub.py
