+++ tests/codemods/codemod_test_cases/func_overload_add_2_no_docstr/doc_stub.pyi
@@ -1,5 +1,6 @@
+# fmt: off
 """
-Overloaded functions
+Overloaded functions, retain existing docstrings
 """
 
 from typing import Literal, overload
