+++ tests/codemods/codemod_test_cases/func_param_update_overwrite/doc_stub.py
@@ -1,9 +1,9 @@
-"""
-Update docstring
-"""
+# fmt: off
+
+from typing import Any
 
 
-def foo(self, iterable) -> None:
+def extend(self, iterable: Sequence[Any], /) -> None:
     """
     Append new elements as contained in *iterable* to the end of
     array, growing it.
