+++ tests/codemods/codemod_test_cases/func_param_update_stdlib/before.py
