+++ tests/codemods/codemod_test_cases/func_param_update_stdlib/expected.py
@@ -18,6 +18,6 @@ if sys.version_info < (3, 12):
         do_handshake: bool = True,
     ) -> SSLSocket:
         """
-        stdlib docs
-        """
+    reference stub
+    """
         ...
