+++ tests/codemods/codemod_test_cases/func_params_update_no_params/before.py
@@ -1,0 +1,12 @@
+# fmt: off
+"""
+simple functions 
+"""
+
+
+def foo(x=1, y=2):
+    ...
+
+
+def bar():
+    ...
