+++ tests/codemods/codemod_test_cases/func_params_update_no_params/doc_stub.py
@@ -1,0 +1,8 @@
+# fmt: off
+"""
+simple functions 
+"""
+
+
+def foo(pin: int, /, limit: int = 100) -> str:
+    ...
