+++ tests/codemods/codemod_test_cases/func_params_update_no_params/expected.py
@@ -1,3 +1,4 @@
+# fmt: off
 """
 simple functions 
 """
