+++ tests/codemods/codemod_test_cases/func_return_update_no_params/doc_stub.py
@@ -1,3 +1,4 @@
+# fmt off
 """
 simple functions 
 """
