+++ tests/codemods/codemod_test_cases/func_return_update_no_params/expected.py
