--- tests/codemods/params_test_cases/func_docstring_add/before.py
@@ -1,11 +1,0 @@
-"""
-Add docstring if not provided
-"""
-
-
-def foo(self, iterable) -> None:
-    # no docstring
-    ...
-
-
-def bar(): ...
