--- tests/codemods/params_test_cases/func_docstring_add/output.py
@@ -1,15 +1,0 @@
-"""
-Add docstring if not provided
-"""
-
-
-def foo(self, iterable) -> None:
-    """
-    Append new elements as contained in *iterable* to the end of
-    array, growing it.
-    """
-    # no docstring
-    ...
-
-
-def bar(): ...
