--- tests/codemods/params_test_cases/func_docstring_no_overwrite/output.py
@@ -1,15 +1,0 @@
-"""
-do not overwrite docstring
-"""
-
-
-def foo(self, iterable) -> None:
-    """
-    original docstring
-    Append new elements as contained in *iterable* to the end of
-    array, growing it.
-    """
-    ...
-
-
-def bar(): ...
