--- tests/codemods/params_test_cases/func_param_update_overwrite/doc_stub.py
@@ -1,12 +1,0 @@
-"""
-simple functions 
-"""
-
-from typing import Any
-
-
-def extend(self, iterable: Sequence[Any], /) -> None:
-    """
-    Append new elements as contained in *iterable* to the end of
-    array, growing it.
-    """
