--- tests/codemods/params_test_cases/func_param_update_overwrite/output.py
@@ -1,15 +1,0 @@
-# fmt:off
-"""
-partially documented function
-"""
-from typing import Any
-
-def extend(self, iterable: Sequence[Any], /) -> None:
-    """
-    Append new elements as contained in *iterable* to the end of
-    array, growing it.
-    """
-    ...
-
-
-def bar(): ...
