--- tests/codemods/params_test_cases/func_param_update_stdlib/output.py
@@ -1,23 +1,0 @@
-# fmt:off
-"""
-conditional function in stdlib stubs
-"""
-import socket
-import sys
-
-if sys.version_info < (3, 12):
-    def wrap_socket(
-        sock: socket.socket,
-        *,
-        server_side: bool = False,
-        key: Incomplete = None,
-        cert: Incomplete = None,
-        cert_reqs: int = 0,
-        cadata: bytes | None = None,
-        server_hostname: str | None = None,
-        do_handshake: bool = True,
-    ) -> SSLSocket:
-        """
-        stdlib docs
-        """
-        ...
